Add files via upload
 Problem statement
Consider a cricket tournament being held involving n teams. Each pair of teams is supposed to
play twice, once at each of the teams' home grounds. Assuming that each team can play only one
game every 2 days, and a schedule for the tournament such that it is completed in the minimum
number of days.